refactor(change_detection): remove circular dependencies
