+++ modules/angular2/src/change_detection/change_detection_jit_generator.es6
@@ -16,9 +16,8 @@ import {
   RECORD_TYPE_KEYED_ACCESS,
   RECORD_TYPE_INVOKE_FORMATTER,
   RECORD_TYPE_STRUCTURAL_CHECK,
-  RECORD_TYPE_INTERPOLATE,
-  ProtoChangeDetector
-  } from './proto_change_detector';
+  RECORD_TYPE_INTERPOLATE
+  } from './proto_record';
 
 /**
  * The code generator takes a list of proto records and creates a function/class
