+++ modules/angular2/src/change_detection/coalesce.js
@@ -1,6 +1,6 @@
 import {isPresent} from 'angular2/src/facade/lang';
 import {List, ListWrapper, Map, MapWrapper} from 'angular2/src/facade/collection';
-import {RECORD_TYPE_SELF, ProtoRecord} from './proto_change_detector';
+import {RECORD_TYPE_SELF, ProtoRecord} from './proto_record';
 
 /**
  * Removes "duplicate" records. It assuming that record evaluation does not
