+++ modules/angular2/src/change_detection/proto_change_detector.js
@@ -31,65 +31,19 @@ import {PipeRegistry} from './pipes/pipe_registry';
 
 import {coalesce} from './coalesce';
 
-export const RECORD_TYPE_SELF = 0;
-export const RECORD_TYPE_CONST = 1;
-export const RECORD_TYPE_PRIMITIVE_OP = 2;
-export const RECORD_TYPE_PROPERTY = 3;
-export const RECORD_TYPE_INVOKE_METHOD = 4;
-export const RECORD_TYPE_INVOKE_CLOSURE = 5;
-export const RECORD_TYPE_KEYED_ACCESS = 6;
-export const RECORD_TYPE_INVOKE_FORMATTER = 7;
-export const RECORD_TYPE_STRUCTURAL_CHECK = 8;
-export const RECORD_TYPE_INTERPOLATE = 9;
-
-export class ProtoRecord {
-  mode:number;
-  name:string;
-  funcOrValue:any;
-  args:List;
-  fixedArgs:List;
-  contextIndex:number;
-  selfIndex:number;
-  bindingMemento:any;
-  directiveMemento:any;
-  lastInBinding:boolean;
-  lastInDirective:boolean;
-  expressionAsString:string;
-
-  constructor(mode:number,
-              name:string,
-              funcOrValue,
-              args:List,
-              fixedArgs:List,
-              contextIndex:number,
-              selfIndex:number,
-              bindingMemento:any,
-              directiveMemento:any,
-              expressionAsString:string,
-              lastInBinding:boolean,
-              lastInDirective:boolean) {
-
-    this.mode = mode;
-    this.name = name;
-    this.funcOrValue = funcOrValue;
-    this.args = args;
-    this.fixedArgs = fixedArgs;
-    this.contextIndex = contextIndex;
-    this.selfIndex = selfIndex;
-    this.bindingMemento = bindingMemento;
-    this.directiveMemento = directiveMemento;
-    this.lastInBinding = lastInBinding;
-    this.lastInDirective = lastInDirective;
-    this.expressionAsString = expressionAsString;
-  }
-
-  isPureFunction():boolean {
-    return this.mode === RECORD_TYPE_INTERPOLATE ||
-      this.mode === RECORD_TYPE_INVOKE_FORMATTER ||
-      this.mode === RECORD_TYPE_PRIMITIVE_OP;
-  }
-}
-
+import {
+  ProtoRecord,
+  RECORD_TYPE_SELF,
+  RECORD_TYPE_PROPERTY,
+  RECORD_TYPE_INVOKE_METHOD,
+  RECORD_TYPE_CONST,
+  RECORD_TYPE_INVOKE_CLOSURE,
+  RECORD_TYPE_PRIMITIVE_OP,
+  RECORD_TYPE_KEYED_ACCESS,
+  RECORD_TYPE_INVOKE_FORMATTER,
+  RECORD_TYPE_STRUCTURAL_CHECK,
+  RECORD_TYPE_INTERPOLATE
+  } from './proto_record';
 
 export class ProtoChangeDetector  {
   addAst(ast:AST, bindingMemento:any, directiveMemento:any = null, structural:boolean = false){}
